fixHT26
