+++ app/src/main/java/com/example/singup_app/presentation/ui/activity/Rest.kt
@@ -13,6 +13,8 @@ import androidx.core.view.WindowInsetsCompat
 import androidx.lifecycle.lifecycleScope
 import com.bumptech.glide.Glide
 import com.example.singup_app.R
+import com.example.singup_app.data.api.retrofitClient.RetrofitClient
+import com.example.singup_app.data.api.server.ServerInt
 import com.google.gson.annotations.SerializedName
 import kotlinx.android.parcel.Parcelize
 import kotlinx.coroutines.Dispatchers
@@ -25,50 +27,22 @@ import retrofit2.http.Path
 
 class Rest : AppCompatActivity() {
 
-    val retrofit = Retrofit.Builder().baseUrl("https://collectionapi.metmuseum.org").addConverterFactory(GsonConverterFactory.create()).build()
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_rest2)
-        val buttonClick =   findViewById<AppCompatButton>(R.id.buttonClick)
+
+        val buttonClick = findViewById<AppCompatButton>(R.id.buttonClick)
         val image = findViewById<AppCompatImageView>(R.id.image)
-        val api = retrofit.create(Server::class.java)
+        val api = RetrofitClient.instance
 
-
-        buttonClick.setOnClickListener{
+        buttonClick.setOnClickListener {
             lifecycleScope.launch(Dispatchers.IO) {
-                withContext(Dispatchers.IO){
-                    val currentBody = api.getSomeObject(45)
+                val currentBody = api.getSomeObject(45)
 
-                    withContext(Dispatchers.Main){
-                        Glide.with(this@Rest).load(currentBody.primaryImage).into(image)
-                    }
+                withContext(Dispatchers.Main) {
+                    Glide.with(this@Rest).load(currentBody.primaryImage).into(image)
                 }
             }
-
-//            lifecycleScope.launch(Dispatchers.IO) {
-//                val responseBody = api.getAllObjects()
-//                Log.d("Response","$responseBody")
-//            }
         }
     }
 }
- interface Server {
-     @GET("/public/collection/v1/objects")
-     suspend fun getAllObjects(): Objects
-
-     @GET("/public/collection/v1/objects/{objectId}")
-     suspend fun getSomeObject(@Path(value = "objectId") objectId: Int): SomeObject
- }
-
-@kotlinx.parcelize.Parcelize
-data class Objects(
-    @SerializedName("total") val total: Int,
-    @SerializedName("objectIds") val objectIds:ArrayList<Int>
-): Parcelable
-
-@kotlinx.parcelize.Parcelize
-data class SomeObject(
-    @SerializedName("objectID") val objectId : Int,
-    @SerializedName("primaryImage") val primaryImage : String
-):Parcelable
